Fix run-time exceptions resulting from Gradle version update
There were various null pointer and file i/o exceptions resulting from
Gradle returning directories, not just files, in the last of changed
items.  The code previously assumed only file changes would be reported,
and therefore acted on the directories as if they were files.

+++ src/main/groovy/com/alliancels/documentation/AssembleDocumentTask.groovy
@@ -35,28 +35,35 @@ class AssembleDocumentTask extends SourceTask {
 
         inputs.outOfDate {
 
-            def relativePath = getRelativePath(it.file)
+            // Only act on files, not directories
+            if (it.file.isFile()) {
+                def relativePath = getRelativePath(it.file)
 
-            def links = new File(linkDir, relativePath).getText('UTF-8')
-            def content = new File(convertedMarkdownDir, relativePath).getText('UTF-8')
+                def links = new File(linkDir, relativePath).getText('UTF-8')
+                def content = new File(convertedMarkdownDir, relativePath).getText('UTF-8')
 
-            // Structure each section like this:
-            // <nav links>
-            // <content>
-            // <nav links>
-            File outputFile = getOutputFile(it.file)
-            String relativeOutputFilePath = outputDir.toPath().relativize(outputFile.toPath()).toString()
-            String completeSection = createCompleteSection(relativeOutputFilePath, links, content, previewEnabled)
-            outputFile.parentFile.mkdirs()
-            outputFile.createNewFile()
-            outputFile.setBytes(completeSection.getBytes("UTF-8"))
+                // Structure each section like this:
+                // <nav links>
+                // <content>
+                // <nav links>
+                File outputFile = getOutputFile(it.file)
+                String relativeOutputFilePath = outputDir.toPath().relativize(outputFile.toPath()).toString()
+                String completeSection = createCompleteSection(relativeOutputFilePath, links, content, previewEnabled)
+                outputFile.parentFile.mkdirs()
+                outputFile.createNewFile()
+                outputFile.setBytes(completeSection.getBytes("UTF-8"))
+            }
         }
 
         inputs.removed {
-            File outputFile = getOutputFile(it.file)
 
-            if (outputFile.exists()) {
-                outputFile.delete()
+			// Only act on files, not directories
+            if (it.file.isFile()) {
+                File outputFile = getOutputFile(it.file)
+
+                if (outputFile.exists()) {
+                    outputFile.delete()
+                }
             }
         }
     }
